make spigot resource updater checker
dont use gson
